feat(calendar): improve responsive design and replace Kumo dialogs
- Replace Kumo Dialog components with custom modal implementations for better control
- Enhance mobile responsiveness with breakpoint-based styling adjustments
- Reorganize calendar header layout to display month/year prominently
- Update formatMonthYear to use short month names for compact display
- Add responsive grid layouts for time labels and day columns

+++ shared/dates.ts
@@ -166,7 +166,7 @@ export function formatShortTime(d: Date): string {
 }
 
 export function formatMonthYear(d: Date): string {
-	return d.toLocaleDateString([], { month: "long", year: "numeric" });
+	return d.toLocaleDateString([], { month: "short", year: "numeric" });
 }
 
 export function formatDayName(d: Date): string {
